+++ src/refactoring/components/cart/coupon/CouponSelector.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Coupon } from "../../../types";
+import { Coupon } from "../../../../types";
 import { CouponOption } from "./CouponOption";
 
 interface CouponSelectorProps {
