+++ src/refactoring/components/cart/product/ProductList.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Product } from "../../../types";
+import { Product } from "../../../../types";
 
 interface ProductListProps {
   products: Product[];
